Separated UI components into files. Moved all ui files into ui module.
