+++ src/ui/mod.rs
@@ -1,0 +1,7 @@
+mod interface;
+mod dropzonearea;
+mod sidebar;
+mod splash;
+mod uitimer;
+
+pub use interface::Interface;
